Refactor userId type in stateManager
The type of the 'userId' parameter in the 'createDefaultState' method of the StateManager class has been updated. It now accepts a string literal 'unknown', instead of the previously used unknown type. This change also reflects in the adAiMessageHandler function where 'unknown' is passed as a string to createDefaultState when currentState is not found.