Added bug report creation functionality
A new function for creating bug reports has been added. This function uses a GraphQL mutation to insert a new bug report into the database. The mutation takes various parameters such as expected outcome, observed outcome, replication steps, and more. The function is exported for use in other modules.

In addition, the route handler for creating bug reports now calls this new function with appropriate arguments. It parses the last message from state messages to get necessary data for creating a bug report and passes it along with game id to the newly created function.